requirements.txt
djangorestframework==3.15.2
Django==5.0.7
djangorestframework-simplejwt==5.3.1
pydantic==2.6.1
pydantic_core==2.16.2
pydantic-settings==2.2.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9